[refactor] Favorite 관련 리팩토링
useCallback, 함수 분리
